+++ cron/hit_stat_test.go
@@ -28,9 +28,8 @@ func TestHitStats(t *testing.T) {
 	}...)
 
 	var stats goatcounter.HitStats
-	total, totalUnique, display, displayUnique, max, more, err := stats.List(
+	total, totalUnique, display, displayUnique, more, err := stats.List(
 		ctx, now.Add(-1*time.Hour), now.Add(1*time.Hour), "", nil, false)
-	_, _, _ = totalUnique, displayUnique, max // TODO
 	if err != nil {
 		t.Fatal(err)
 	}