+++ handlers/backend.go
@@ -145,7 +145,6 @@ func (h backend) Mount(r chi.Router, db zdb.DB) {
 			ap.Get("/systems", zhttp.Wrap(h.systems))
 			ap.Get("/sizes", zhttp.Wrap(h.sizes))
 			ap.Get("/locations", zhttp.Wrap(h.locations))
-			ap.Get("/toprefs", zhttp.Wrap(h.topRefs))
 			ap.Get("/pages-by-ref", zhttp.Wrap(h.pagesByRef))
 		}
 		{
@@ -325,7 +324,6 @@ func (h backend) index(w http.ResponseWriter, r *http.Request) error {
 		pages                           goatcounter.HitStats
 		total, totalDisplay             int
 		totalUnique, totalUniqueDisplay int
-		max                             int
 		morePages                       bool
 		pagesErr                        error
 	)
@@ -334,8 +332,21 @@ func (h backend) index(w http.ResponseWriter, r *http.Request) error {
 		defer zlog.Recover()
 		defer wg.Done()
 
-		total, totalUnique, totalDisplay, totalUniqueDisplay, max, morePages, pagesErr = pages.List(r.Context(), start, end, filter, nil, daily)
-		//l = l.Since("pages.List")
+		total, totalUnique, totalDisplay, totalUniqueDisplay, morePages, pagesErr = pages.List(
+			r.Context(), start, end, filter, nil, daily)
+	}()
+
+	var (
+		totalPages goatcounter.HitStat
+		max        int
+		totalErr   error
+	)
+	wg.Add(1)
+	go func() {
+		defer zlog.Recover()
+		defer wg.Done()
+
+		max, totalErr = totalPages.Totals(r.Context(), start, end, filter, daily)
 	}()
 
 	var browsers goatcounter.Stats
@@ -343,21 +354,18 @@ func (h backend) index(w http.ResponseWriter, r *http.Request) error {
 	if err != nil {
 		return err
 	}
-	l = l.Since("browsers.List")
 
 	var systems goatcounter.Stats
 	totalSystems, err := systems.ListSystems(r.Context(), start, end)
 	if err != nil {
 		return err
 	}
-	l = l.Since("systems.List")
 
 	var sizeStat goatcounter.Stats
 	totalSize, err := sizeStat.ListSizes(r.Context(), start, end)
 	if err != nil {
 		return err
 	}
-	l = l.Since("sizeStat.ListSizes")
 
 	var locStat goatcounter.Stats
 	totalLoc, err := locStat.ListLocations(r.Context(), start, end)
@@ -365,25 +373,22 @@ func (h backend) index(w http.ResponseWriter, r *http.Request) error {
 		return err
 	}
 	showMoreLoc := len(locStat) > 0 && float32(locStat[len(locStat)-1].Count)/float32(totalLoc)*100 < 3.0
-	l = l.Since("locStat.List")
-
-	var topRefs goatcounter.Stats
-	totalTopRefs, showMoreRefs, err := topRefs.ListRefs(r.Context(), start, end, 10, 0)
-	if err != nil {
-		return err
-	}
-	l = l.Since("topRefs.List")
 
 	// Add refers.
 	sr := r.URL.Query().Get("showrefs")
-	var refs goatcounter.HitStats
-	var moreRefs bool
+	var (
+		refs     goatcounter.HitStats
+		moreRefs bool
+	)
 	if sr != "" {
-		moreRefs, err = refs.ListRefs(r.Context(), sr, start, end, 0)
+		if sr == goatcounter.PathTotals {
+			moreRefs, err = refs.ListAllRefs(r.Context(), start, end, 0)
+		} else {
+			moreRefs, err = refs.ListRefs(r.Context(), sr, start, end, 0)
+		}
 		if err != nil {
 			return err
 		}
-		l = l.Since("refs.ListRefs")
 	}
 
 	subs, err := site.ListSubs(r.Context())
@@ -403,6 +408,9 @@ func (h backend) index(w http.ResponseWriter, r *http.Request) error {
 	if pagesErr != nil {
 		return pagesErr
 	}
+	if totalErr != nil {
+		return totalErr
+	}
 
 	l = startl.Since("get data")
 
@@ -417,6 +425,7 @@ func (h backend) index(w http.ResponseWriter, r *http.Request) error {
 		PeriodEnd          time.Time
 		Filter             string
 		Pages              goatcounter.HitStats
+		TotalPages         goatcounter.HitStat
 		MorePages          bool
 		Refs               goatcounter.HitStats
 		MoreRefs           bool
@@ -434,41 +443,16 @@ func (h backend) index(w http.ResponseWriter, r *http.Request) error {
 		LocationStat       goatcounter.Stats
 		TotalLocation      int
 		ShowMoreLocations  bool
-		TopRefs            goatcounter.Stats
-		TotalTopRefs       int
-		ShowMoreRefs       bool
 		Daily              bool
 		ForcedDaily        bool
 		Max                int
 	}{newGlobals(w, r), cd, sr, r.URL.Query().Get("hl-period"), start, end,
-		filter, pages, morePages, refs, moreRefs, total, totalUnique,
-		totalDisplay, totalUniqueDisplay, browsers, totalBrowsers, systems,
-		totalSystems, subs, sizeStat, totalSize, locStat, totalLoc, showMoreLoc,
-		topRefs, totalTopRefs, showMoreRefs, daily, forcedDaily, max})
+		filter, pages, totalPages, morePages, refs, moreRefs, total,
+		totalUnique, totalDisplay, totalUniqueDisplay, browsers, totalBrowsers,
+		systems, totalSystems, subs, sizeStat, totalSize, locStat, totalLoc,
+		showMoreLoc, daily, forcedDaily, max})
 	l.Since("zhttp.Template")
 	return x
-}
-
-func (h backend) topRefs(w http.ResponseWriter, r *http.Request) error {
-	start, end, err := getPeriod(w, r, goatcounter.MustGetSite(r.Context()))
-	if err != nil {
-		return err
-	}
-
-	var refs goatcounter.Stats
-	o, _ := strconv.ParseInt(r.URL.Query().Get("offset"), 10, 64)
-	total, hasMore, err := refs.ListRefs(r.Context(), start, end, 10, int(o))
-	if err != nil {
-		return err
-	}
-
-	t, _ := strconv.ParseInt(r.URL.Query().Get("total"), 10, 64)
-	tpl := goatcounter.HorizontalChart(r.Context(), refs, total, int(t), 0, true, false)
-
-	return zhttp.JSON(w, map[string]interface{}{
-		"html":     string(tpl),
-		"has_more": hasMore,
-	})
 }
 
 func (h backend) pagesByRef(w http.ResponseWriter, r *http.Request) error {
@@ -505,10 +489,22 @@ func (h backend) refs(w http.ResponseWriter, r *http.Request) error {
 		offset = int(o2)
 	}
 
-	var refs goatcounter.HitStats
-	more, err := refs.ListRefs(r.Context(), r.URL.Query().Get("showrefs"), start, end, offset)
-	if err != nil {
-		return err
+	showRefs := r.URL.Query().Get("showrefs")
+
+	var (
+		refs goatcounter.HitStats
+		more bool
+	)
+	if showRefs == goatcounter.PathTotals {
+		more, err = refs.ListAllRefs(r.Context(), start, end, offset)
+		if err != nil {
+			return err
+		}
+	} else {
+		more, err = refs.ListRefs(r.Context(), showRefs, start, end, offset)
+		if err != nil {
+			return err
+		}
 	}
 
 	tpl, err := zhttp.ExecuteTpl("_backend_refs.gohtml", map[string]interface{}{
@@ -606,15 +602,59 @@ func (h backend) locations(w http.ResponseWriter, r *http.Request) error {
 func (h backend) pages(w http.ResponseWriter, r *http.Request) error {
 	site := goatcounter.MustGetSite(r.Context())
 
+	exclude := r.URL.Query().Get("exclude")
+	filter := r.URL.Query().Get("filter")
 	start, end, err := getPeriod(w, r, site)
 	if err != nil {
 		return err
 	}
 	daily, forcedDaily := getDaily(r, start, end)
+	m, err := strconv.ParseInt(r.URL.Query().Get("max"), 10, 64)
+	if err != nil {
+		return err
+	}
+	max := int(m)
+
+	// Load new totals unless this is for pagination.
+	var (
+		wg         sync.WaitGroup
+		totalTpl   []byte
+		totalPages goatcounter.HitStat
+		totalErr   error
+	)
+	if exclude == "" {
+		wg.Add(1)
+		go func() {
+			defer zlog.Recover()
+			defer wg.Done()
+
+			max, totalErr = totalPages.Totals(r.Context(), start, end, filter, daily)
+			if totalErr != nil {
+				return
+			}
+
+			totalTpl, totalErr = zhttp.ExecuteTpl("_backend_totals.gohtml", struct {
+				Context     context.Context
+				Site        *goatcounter.Site
+				PeriodStart time.Time
+				PeriodEnd   time.Time
+				TotalPages  goatcounter.HitStat
+				Daily       bool
+				Max         int
+
+				// Dummy values so template won't error out.
+				TotalUniqueDisplay int
+				TotalHitsDisplay   int
+				Refs               bool
+				ShowRefs           string
+			}{r.Context(), site, start, end, totalPages, daily, max,
+				0, 0, false, ""})
+		}()
+	}
 
 	var pages goatcounter.HitStats
-	totalHits, totalUnique, totalDisplay, totalUniqueDisplay, max, more, err := pages.List(r.Context(), start, end,
-		r.URL.Query().Get("filter"), strings.Split(r.URL.Query().Get("exclude"), ","), daily)
+	totalHits, totalUnique, totalDisplay, totalUniqueDisplay, more, err := pages.List(
+		r.Context(), start, end, filter, strings.Split(exclude, ","), daily)
 	if err != nil {
 		return err
 	}
@@ -632,8 +672,8 @@ func (h backend) pages(w http.ResponseWriter, r *http.Request) error {
 		// Dummy values so template won't error out.
 		Refs     bool
 		ShowRefs string
-	}{r.Context(), pages, goatcounter.MustGetSite(r.Context()), start, end,
-		daily, forcedDaily, max, false, ""})
+	}{r.Context(), pages, site, start, end,
+		daily, forcedDaily, int(max), false, ""})
 	if err != nil {
 		return err
 	}
@@ -643,13 +683,20 @@ func (h backend) pages(w http.ResponseWriter, r *http.Request) error {
 		paths[i] = pages[i].Path
 	}
 
+	wg.Wait()
+	if totalErr != nil {
+		return totalErr
+	}
+
 	return zhttp.JSON(w, map[string]interface{}{
 		"rows":                 string(tpl),
+		"totals":               string(totalTpl),
 		"paths":                paths,
 		"total_hits":           totalHits,
 		"total_display":        totalDisplay,
 		"total_unique":         totalUnique,
 		"total_unique_display": totalUniqueDisplay,
+		"max":                  max,
 		"more":                 more,
 	})
 }