+++ pack/pack.go
@@ -12264,7 +12264,7 @@ http://nicolasgallagher.com/micro-clearfix-hack/
 		;[report_errors, period_select, load_refs, tooltip, paginate_paths,
 			paginate_refs, hchart_detail, settings_tabs, paginate_locations,
 			billing_subscribe, setup_datepicker, filter_paths, add_ip, fill_tz,
-			paginate_toprefs, draw_chart, tsort,
+			draw_chart, bind_scale, tsort,
 		].forEach(function(f) { f.call() })
 	});
 
@@ -12297,13 +12297,10 @@ http://nicolasgallagher.com/micro-clearfix-hack/
 		});
 	}
 
-	// Replace the "height:" style with a background gradient and set the height
-	// to 100%.
-	//
-	// This way you can still hover the entire height.
-	var draw_chart = function() {
+	// Bind the Y-axis scale actions.
+	var bind_scale = function() {
 		var redraw = () => {
-			if ($('#scale').val() === $('.count-list-pages').attr('data-scale'))
+			if ($('#scale').val() === get_max())
 				return $('#scale').removeClass('value')
 
 			$('#scale').addClass('value')
@@ -12323,14 +12320,27 @@ http://nicolasgallagher.com/micro-clearfix-hack/
 				t = setTimeout(redraw, 300)
 			})
 
-		$('#scale-reset').on('click', (e) => {
+		$('#scale-half').on('click', (e) => {
 			clearTimeout(t)
 			e.preventDefault()
-			$('#scale').val($('.count-list-pages').attr('data-max'))
+			$('#scale').val(Math.max(10, Math.ceil(parseInt($('#scale').val(), 10) / 2)))
 			redraw()
 		})
 
-		var scale = parseInt($('#scale').val(), 10) / parseInt($('.count-list-pages').attr('data-max'), 10)
+		$('#scale-reset').on('click', (e) => {
+			clearTimeout(t)
+			e.preventDefault()
+			$('#scale').val(get_max())
+			redraw()
+		})
+	}
+
+	// Replace the "height:" style with a background gradient and set the height
+	// to 100%.
+	//
+	// This way you can still hover the entire height.
+	var draw_chart = function() {
+		var scale = parseInt($('#scale').val(), 10) / parseInt(get_max(), 10)
 		$('.chart-bar').each(function(i, chart) {
 			if (chart.dataset.done === 't')
 				return
@@ -12419,6 +12429,11 @@ http://nicolasgallagher.com/micro-clearfix-hack/
 		});
 	};
 
+	// Get the original y-axis max.
+	var get_max = function() {
+		return $('.count-list-pages').attr('data-max')
+	}
+
 	// Reload the path list when typing in the filter input, so the user won't
 	// have to press "enter".
 	var filter_paths = function() {
@@ -12436,7 +12451,11 @@ http://nicolasgallagher.com/micro-clearfix-hack/
 				$(e.target).after(loading)
 				jQuery.ajax({
 					url:     '/pages',
-					data:    append_period({filter: filter}),
+					data:    append_period({
+						filter: filter,
+						daily:  $('#daily').is(':checked'),
+						max:    get_max(),
+					}),
 					success: function(data) {
 						update_pages(data, true)
 						loading.remove()
@@ -12461,8 +12480,9 @@ http://nicolasgallagher.com/micro-clearfix-hack/
 					url:  '/pages',
 					data: append_period({
 						filter:  $('#filter-paths').val(),
-						daily:   $('#daily').is(':selected'),
+						daily:   $('#daily').is(':checked'),
 						exclude: $('.count-list-pages >tbody >tr').toArray().map((e) => e.id).join(','),
+						max:     get_max(),
 					}),
 					success: function(data) {
 						update_pages(data, false)
@@ -12475,10 +12495,15 @@ http://nicolasgallagher.com/micro-clearfix-hack/
 
 	// Update the page list from ajax request on pagination/filter.
 	var update_pages = function(data, from_filter) {
-		if (from_filter)
-			$('.pages-list .count-list-pages > tbody').html(data.rows);
+		if (from_filter) {
+			$('.count-list-pages').attr('data-max', data.max)
+			$('#scale').val(data.max)
+
+			$('.pages-list .count-list-pages > tbody.totals').replaceWith(data.totals)
+			$('.pages-list .count-list-pages > tbody.pages').html(data.rows)
+		}
 		else
-			$('.pages-list .count-list-pages > tbody').append(data.rows);
+			$('.pages-list .count-list-pages > tbody.pages').append(data.rows)
 
 		highlight_filter($('#filter-paths').val())
 		$('.pages-list .load-more').css('display', data.more ? 'inline' : 'none')
@@ -12509,7 +12534,9 @@ http://nicolasgallagher.com/micro-clearfix-hack/
 	var highlight_filter = function(s) {
 		if (s === '')
 			return;
-		$('.pages-list .count-list-pages > tbody').find('.rlink, .page-title:not(.no-title)').each(function(_, elem) {
+		$('.pages-list .count-list-pages > tbody.pages').find('.rlink, .page-title:not(.no-title)').each(function(_, elem) {
+			if ($(elem).find('b').length)  // Don't apply twice after pagination
+				return
 			elem.innerHTML = elem.innerHTML.replace(new RegExp('' + quote_re(s) + '', 'gi'), '<b>$&</b>');
 		});
 	};
@@ -12579,30 +12606,6 @@ http://nicolasgallagher.com/micro-clearfix-hack/
 				},
 			});
 		});
-	};
-
-	// Paginate the top ref list.
-	var paginate_toprefs = function() {
-		$('.top-refs-chart .show-more').on('click', function(e) {
-			e.preventDefault();
-
-			var bar = $(this).parent().find('.chart-hbar:first')
-			var done = paginate_button($(this), () => {
-				jQuery.ajax({
-					url: '/toprefs',
-					data: append_period({
-						offset: $('.top-refs-chart [data-detail] > a').length,
-						total:  $('.total-hits').text().replace(/[^\d]/, ''),
-					}),
-					success: function(data) {
-						bar.append(data.html)
-						if (!data.has_more)
-							$('.top-refs-chart .show-more').remove()
-						done()
-					},
-				})
-			})
-		})
 	};
 
 	// Paginate the location chart.
@@ -12792,7 +12795,7 @@ http://nicolasgallagher.com/micro-clearfix-hack/
 
 	// Load references as an AJAX request.
 	var load_refs = function() {
-		$('.count-list-pages').on('click', '.rlink', function(e) {
+		$('.count-list-pages').on('click', '.load-refs', function(e) {
 			e.preventDefault()
 
 			var params = split_query(location.search),
@@ -13334,7 +13337,7 @@ form .err  { color: red; display: block; }
 
 .count-list td:first-child {  /* Count */
 	text-align: right;
-	width: 5em;
+	width: 5rem;
 }
 .count-list td[colspan="3"] {  /* "nothing to display" */
 	text-align: left;
@@ -13342,21 +13345,29 @@ form .err  { color: red; display: block; }
 }
 
 .count-list td:nth-child(2) {  /* Path */
-	width: 20em;
+	width: 20rem;
 }
 
 .count-list.count-list-refs td:nth-child(1) {
 	text-align: right;
-	width: 4em;
+	width: 4rem;
 }
 .count-list.count-list-refs td:nth-child(2) {
 	text-align: right;
-	width: 4em;
+	width: 4rem;
 }
 
 .count-list.count-list-refs td:nth-child(3) {
 	width: auto;
 	word-break: break-all; /* don't make it wider for very long urls */
+}
+
+/* Totals */
+.count-list .totals                 { background-color: #f7f7f7; border-bottom: 1px solid #999; }
+.count-list .pages::before { content: ''; display: block; height: 1rem; } /* Hack to add margin to tbody */
+.count-list .totals .load-refs.desktop {
+	float: right;
+	margin-right: 1em;
 }
 
 .label-event { background-color: #f6f3da; border-radius: 1em; padding: .1em .3em; }
@@ -13379,8 +13390,11 @@ form .err  { color: red; display: block; }
 	background-color: inherit;
 }
 
-.count-list tr:target > td:nth-child(2) .rlink,
-.count-list tr.target > td:nth-child(2) .rlink {
+/* Border doesn't affect layout. */
+.count-list .load-refs { border-bottom: 4px solid transparent; margin-bottom: -4px; }
+
+.count-list tr:target > td:nth-child(2) .load-refs,
+.count-list tr.target > td:nth-child(2) .load-refs {
 	font-weight: bold;
 	border-bottom: 4px solid yellow;
 }
@@ -13634,25 +13648,19 @@ noscript p { margin: .5em; }
 header h2 { border-bottom: 0; display: inline; }
 header.h2 { border-bottom: 1px solid #252525; padding-bottom: .2em; margin: 1em 0; }
 
-.header-pages         { display: flex; font-size: .9rem; }
+.header-pages         { display: flex; justify-content: space-between; font-size: .9rem; }
 .header-pages h2      { margin: 0; margin-right: 1em; display: none; }
 .header-pages span    { margin-left: 0; }
 .header-pages .totals { flex-grow: 1; }
 .header-pages input#filter-paths,
 .header-pages input#scale,
-.header-pages select#display { padding: .2em; margin-right: 1em; }
+.header-pages select#display { padding: .2em; }
 .header-pages input.value    { background-color: yellow; }
 .header-pages .totals-small  { display: none; }
 
-@media (max-width: 36rem) {
-	.header-pages input#filter-paths { max-width: 10em; }
-	.header-pages .totals       { display: none; }
-	.header-pages .totals-small { display: block; }
-}
+.header-pages input#scale { width: 6em; margin-right: .2em; }
+.header-pages .scale-wrap { }
 
-.header-pages input#scale { width: 6em; }
-/* Don't wrap on smaller screens; prefer counts to wrap instead. */
-.header-pages .scale-wrap { white-space: no-wrap; }
 
 h3 + h4 { margin-top: .3em; }
 
@@ -13669,6 +13677,7 @@ h3 + h4 { margin-top: .3em; }
   50%  { content: ".."; }
   100% { content: "..."; }
 }
+/* TODO: See if we can have it sit in the margin */
 a.loading        { color: #777; }
 a.loading:hover  { color: #777; text-decoration: none; }
 a.loading::after { content: ""; animation: loading 500ms linear infinite; }
@@ -13679,6 +13688,13 @@ small.loading::after { content: ""; animation: loading 500ms linear infinite; }
 /* Filter */
 .filter-wrap                 { position: relative; }
 .filter-wrap .loading::after { position: absolute; right: 1.2em; content: ""; animation: loading 500ms linear infinite; }
+.header-pages .filter-wrap   { margin-left: 5.5rem; }
+@media (max-width: 32rem) {
+	.header-pages .filter-wrap       { margin-left: .2rem; }
+}
+@media (max-width: 28rem) {
+	.header-pages input#filter-paths { max-width: 10em; }
+}
 `),
 }
 
@@ -14833,14 +14849,14 @@ var Templates = map[string][]byte{
 			<span title="Pageviews" class="views">{{nformat $h.Count $.Site}}</span><br>
 		</td>
 		<td class="hide-mobile">
-			<a class="rlink" title="{{$h.Path}}" href="?showrefs={{$h.Path}}&period-start={{tformat $.Site $.PeriodStart ""}}&period-end={{tformat $.Site $.PeriodEnd ""}}#{{$h.Path}}">{{$h.Path}}</a><br>
+			<a class="load-refs rlink" title="{{$h.Path}}" href="?showrefs={{$h.Path}}&period-start={{tformat $.Site $.PeriodStart ""}}&period-end={{tformat $.Site $.PeriodEnd ""}}#{{$h.Path}}">{{$h.Path}}</a><br>
 			<small class="page-title {{if not $h.Title}}no-title{{end}}">{{if $h.Title}}{{$h.Title}}{{else}}<em>(no title)</em>{{end}}</small>
 			{{if $h.Event}}<sup class="label-event">event</sup>{{end}}
 			{{if and $.Site.LinkDomain (not $h.Event)}}<sup><a class="go" target="_blank" rel="noopener" href="https://{{$.Site.LinkDomain}}{{$h.Path}}">go</a></sup>{{end}}
 		</td>
 		<td>
 			<div class="show-mobile">
-				<a class="rlink" title="{{$h.Path}}" href="?showrefs={{$h.Path}}&period-start={{tformat $.Site $.PeriodStart ""}}&period-end={{tformat $.Site $.PeriodEnd ""}}#{{$h.Path}}">{{$h.Path}}</a>
+				<a class="load-refs rlink" title="{{$h.Path}}" href="?showrefs={{$h.Path}}&period-start={{tformat $.Site $.PeriodStart ""}}&period-end={{tformat $.Site $.PeriodEnd ""}}#{{$h.Path}}">{{$h.Path}}</a>
 				<small class="page-title {{if not $h.Title}}no-title{{end}}">| {{if $h.Title}}{{$h.Title}}{{else}}<em>(no title)</em>{{end}}</small>
 				{{if $h.Event}}<sup class="label-event">event</sup>{{end}}
 				{{if and $.Site.LinkDomain (not $h.Event)}}<sup><a class="go" target="_blank" rel="noopener" href="https://{{$.Site.LinkDomain}}{{$h.Path}}">go</a></sup>{{end}}
@@ -15525,6 +15541,39 @@ want to modify that in JavaScript; you can use <code>goatcounter.endpoint</code>
 	<div class="page">
 	{{- if .Flash}}<div class="flash flash-{{.Flash.Level}}">{{.Flash.Message}}</div>{{end -}}
 `),
+	"tpl/_backend_totals.gohtml": []byte(`<tbody class="totals"><tr id="TOTAL "{{if eq "TOTAL " $.ShowRefs}}class="target"{{end}}>
+	<td>
+		<span title="Visits">{{nformat .TotalPages.CountUnique $.Site}}</span><br>
+		<span title="Pageviews" class="views">{{nformat .TotalPages.Count $.Site}}</span><br>
+	</td>
+	<td class="hide-mobile">
+		<strong title="Totals for all the pages">Totals</strong>
+		<a class="load-refs desktop" href="?showrefs=TOTAL%20&period-start={{tformat $.Site $.PeriodStart ""}}&period-end={{tformat $.Site $.PeriodEnd ""}}#TOTAL%20">Top referrers</a><br>
+		<small>
+			Displaying <span class="total-unique-display">{{nformat .TotalUniqueDisplay $.Site}}</span> visits;
+			<span class="views"><span class="total-display">{{nformat .TotalHitsDisplay $.Site}}</span> pageviews</span>
+		</small>
+	</td>
+	<td>
+		<div class="show-mobile">
+			<strong>Total</strong> |
+			<small>
+				Displaying <span class="total-unique-display" title="Visits">{{nformat .TotalUniqueDisplay $.Site}}</span> visits;
+				<span class="views"><span class="total-display">{{nformat .TotalHitsDisplay $.Site}}</span> pageviews</span>
+			</small> |
+			<a class="load-refs" href="?showrefs=TOTAL%20&period-start={{tformat $.Site $.PeriodStart ""}}&period-end={{tformat $.Site $.PeriodEnd ""}}#TOTAL%20">Top referrers</a>
+		</div>
+		<div class="chart chart-bar chart-totals">
+			<span class="half"></span>
+			{{bar_chart .Context .TotalPages.Stats .Max .Daily}}
+		</div>
+		<div class="refs">{{if and $.Refs (eq $.ShowRefs "TOTAL ")}}
+			{{template "_backend_refs.gohtml" map "Refs" $.Refs "Site" $.Site}}
+			{{if $.MoreRefs}}<a href="#_", class="load-more-refs">Show more</a>{{end}}
+		{{end}}</div>
+	</td>
+</tr></tbody>
+`),
 	"tpl/_bottom.gohtml": []byte(`		<script crossorigin="anonymous" src="{{.Static}}/imgzoom.js?v={{.Version}}"></script>
 		<script crossorigin="anonymous" src="{{.Static}}/script.js?v={{.Version}}"></script>
 	</div> {{/* .page */}}
@@ -15724,40 +15773,24 @@ Martin
 	<div class="pages-list {{if .Daily}}pages-list-daily{{end}}">
 		<header class="h2 header-pages">
 			<h2>Paths</h2>
-			<span class="totals">
-				Displaying <span class="total-unique-display">{{nformat .TotalUniqueDisplay $.Site}}</span>
-				out of <span class="total-unique">{{nformat .TotalUniqueHits $.Site}}</span>
-				visits;
-				<br class="show-mobile">
-				<span class="views">
-					<span class="total-display">{{nformat .TotalHitsDisplay $.Site}}</span> out of
-					<span class="total-hits">{{nformat .TotalHits $.Site}}</span> pageviews
-				</span>
-			</span>
-
-			<span class="totals totals-small">
-				<span class="total-unique-display">{{nformat .TotalUniqueDisplay $.Site}}</span>/<span class="total-unique">{{nformat .TotalUniqueHits $.Site}}</span> visits
-				<br class="show-mobile">
-				<span class="views">
-					<span class="total-display">{{nformat .TotalHitsDisplay $.Site}}</span>/<span class="total-hits">{{nformat .TotalHits $.Site}}</span> views
-				</span>
-			</span>
-
-			<div class="scale-wrap">
-				<a href="#" id="scale-reset" title="Reset Y-axis scale to the default value">reset</a>
-				<input type="number"autocomplete="off" name="scale" id="scale" value="{{.Max}}"
-					placeholder="Scale" title="Set the Y-axis scale">
-			</div>
-
 			<div class="filter-wrap">
 				<input type="text" autocomplete="off" name="filter" value="{{.Filter}}" id="filter-paths"
 					placeholder="Filter paths" title="Filter the list of paths; matched case-insensitive on path and title"
 					{{if .Filter}}class="value"{{end}}>
 			</div>
+
+			<div class="scale-wrap">
+				<label for="scale">Y-axis scale</label>
+				<input type="number" autocomplete="off" name="scale" id="scale" value="{{.Max}}"
+					placeholder="Scale" title="Set the Y-axis scale">
+				<a href="#" id="scale-reset" title="Reset Y-axis scale to the default value">reset</a>;
+				<a href="#" id="scale-half" title="Set Y-axis scale to half the current value">half</a>
+			</div>
 		</header>
 
 		<table class="count-list count-list-pages" data-max="{{.Max}}" data-scale="{{.Max}}">
-			<tbody>{{template "_backend_pages.gohtml" .}}</tbody>
+			{{template "_backend_totals.gohtml" .}}
+			<tbody class="pages">{{template "_backend_pages.gohtml" .}}</tbody>
 		</table>
 
 		<a href="#" class="load-more" {{if not .MorePages}}style="display: none"{{end}}>Show more</a>
@@ -15806,17 +15839,6 @@ Martin
 				<div class="chart-hbar">{{horizontal_chart .Context .LocationStat .TotalLocation 0 3 false true}}</div>
 			</div>
 			{{if .ShowMoreLocations}}<a href="#" class="show-all">Show all</a>{{end}}
-		{{end}}
-	</div>
-	<div class="top-refs-chart">
-		<h2>Top referrers</h2>
-		{{if eq .TotalHits 0}}
-			<em>Nothing to display</em>
-		{{else}}
-			<div class="hchart-wrap">
-				<div class="chart-hbar" data-detail="/pages-by-ref">{{horizontal_chart .Context .TopRefs .TotalTopRefs 0 0 true false}}</div>
-			</div>
-			{{if .ShowMoreRefs}}<a href="#" class="show-more">Show more</a>{{end}}
 		{{end}}
 	</div>
 </div>